+++ src/main/java/com/example/demo/dto/CommentRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class CommentRequest 
 {
 	private String comment;
-	private int postId;
-	private int userId;
+	private long postId;
+	private long userId;
 
 }
